Open the range editor as a sidebar through Umbraco's modal manager
The inline uui-modal-dialog rendered inside the property editor's own template,
so it appeared below the grid rather than over it and sat outside Umbraco's
focus and layering handling.

The range editor is now a registered modal extension opened with umbOpenModal,
configured as a sidebar so it slides in from the right like a content picker.
Cancelling rejects the promise and leaves the day untouched; saving and removing
both resolve with the day's full set of ranges.
